succeed to download via HTTP partially 1 fix the bug in len and bw calculation (overflow) 2 fix the bug in task scheduling (task index error)

+++ app/src/main/java/edu/pdx/cs410/wifi/direct/file/transfer/trans/BwMetric.java
@@ -54,7 +54,7 @@ public class BwMetric {
         lastTime = currenntTime;
         alreadyTime += timeSpan;
         if (alreadyTime > 0) {
-            bw = 1000 * alreadyBytes / (alreadyTime * 1024);
+            bw = (long)((float)1000 * ((float)alreadyBytes / (float)(alreadyTime * 1024)));
         } else {
             bw = 0;
         }
